backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.25
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Document Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
pypdf==3.17.4
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.1.4

# OCR & Vision
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.2.0

# AI & ML
google-generativeai>=0.7.0
openai>=1.0.0
requests>=2.31.0
scikit-learn==1.4.0
xgboost==2.0.3
numpy==1.26.3
shap==0.44.0
matplotlib==3.8.2

# Web Scraping
beautifulsoup4==4.12.3
selenium==4.16.0

# Graph Analysis
networkx>=3.2

# Utilities
aiofiles==23.2.1
httpx==0.26.0
joblib>=1.3.0
